In Dr Schofld's own words
this patch fixes the build problem in opensim-libs/libsl1550 with nant/mcs barfing about the missing message_template.msg --- the real bug is in the mapgenerator. however, since we don't need to regenerate the _Packet_.cs with the current grid, we can just disable the dependency on "map" in libsl1550/libsecondlife/libsecondlife.build --- which is what this patch does.


git-svn-id: file:///var/svn/opensim-libs@32 7c4940e8-6ff7-4914-b775-378afef7cd2e