Remove the CS code generation files originally duplicated from OpenSim proper.
These were probably present for testing purposes, but various regression tests are now carried out in OpenSimulator directly.
These versions of the files were also not maintained and so would not work anyway.
Leaving the tests for now since they may still be useful eventually in opensim proper (but are also probably too reliant on exact formatting to be particularly useful).